requirements.txt
cryptography==41.0.7
pycryptodome==3.19.0
qrcode==7.4.2
Pillow==10.1.0
colorama==0.4.6
tabulate==0.9.0